The necessary adjustments are made and the CSS is created.
All the main details of both the HTML code and the CSS are finished. There are still some details missing, but the most important parts are already there. We still need to improve the page in general and settle on a single CSS base for everything relevant that is needed, such as images, text, titles, etc.